refactor(timeblock): statistics-service をドメイン単位の service に分割する
statistics-service.ts（763行）を facade 化し、実装を責務ごとに移動:

- statistics-general-service.ts: タグ別統計 / 時間帯分布 / トレンド
- statistics-kpi-service.ts: 見積もり精度 / 空白率
- statistics-summary-service.ts: streak / KPI サマリー / Time P/L / Review panel 統合
- statistics-tag-dashboard-service.ts: タグ詳細ダッシュボード
- statistics-fetchers.ts: 行取得（fetchRecords/fetchPlans/fetchTagsById 等、共有）
- statistics-row-builders.ts: 行組み立て（buildTagPL 等、共有）

ロジックは書き換えずに移動のみ。既存 router からの呼び出しシグネチャは不変。
既存 statistics-service.test.ts は facade 経由のまま無変更で green。全ファイル 400 行未満。

Refs #1528
