+++ apps/product/src/features/timeblock/server/statistics-service.ts
@@ -1,43 +1,5 @@
 import 'server-only';
 
-import { formatInTimeZone, fromZonedTime } from 'date-fns-tz';
-
-import { databaseTables } from '@/lib/database';
-import { getUserTimezone } from '@/lib/server/user-timezone-cache';
-
-import {
-  aggregateDayOfWeekDistribution,
-  aggregateHourlyDistribution,
-  aggregateMonthlyTrend,
-  aggregatePlanRecordEstimationAccuracy,
-  aggregateTagStats,
-  type EstimationAccuracyDbRow,
-  type EstimationAccuracyTagLookup,
-  getMonthlyStartDate,
-  transformEstimationAccuracy,
-} from '../domain';
-import {
-  buildTagDashboard,
-  type TagDashboardTagRow,
-  type TagDashboardTimeblockRow,
-} from '../domain/tag-dashboard';
-
-import { transformStatsOverviewResponse } from './statistics-overview-transform';
-import {
-  computeAvailableMinutesInclusive,
-  computeBlankRate,
-  computeContextSwitches,
-  groupEnergyMap,
-  groupHoursByDay,
-  groupHoursByMonth,
-  groupMinutesByDow,
-  groupMinutesByHour,
-  minutesBetween,
-} from './statistics-service-grouping';
-import type { StatsPageData, TimePLResponse } from './statistics-shared';
-import { transformTimeByTagResponse } from './statistics-time-by-tag-transform';
-import type { ServiceSupabaseClient } from './types';
-
 /**
  * Step 4: 統計 TS service。
  *
@@ -47,186 +9,70 @@ import type { ServiceSupabaseClient } from './types';
  * **dormant**: このクラスは router から未接続。既存 RPC ベースの router
  * (`statistics-general-router.ts` 等) は現状維持のまま、この service は Step 8 の
  * カットオーバーで router 内部の実装として丸ごと差し替えられる前提のシグネチャを持つ。
+ *
+ * 公開 API は facade（このファイル）。実装はドメイン単位の service に分割されている:
+ * - General（タグ別統計・時間帯分布・トレンド）: statistics-general-service.ts
+ * - KPI（見積もり精度・空白率）: statistics-kpi-service.ts
+ * - Summary（streak / KPI サマリー / Time P/L / Review panel 統合データ）: statistics-summary-service.ts
+ * - タグ詳細ダッシュボード: statistics-tag-dashboard-service.ts
+ * - 行取得: statistics-fetchers.ts / 行組み立て: statistics-row-builders.ts
  */
 
-interface StatPlanRow {
-  id: string;
-  tag_id: string | null;
-  start_at: string;
-  end_at: string;
-}
-
-interface StatRecordRow {
-  id: string;
-  tag_id: string | null;
-  plan_id: string | null;
-  source: string;
-  start_at: string;
-  end_at: string;
-}
-
-interface TagLookupRow {
-  id: string;
-  name: string;
-  color: string | null;
-  icon: string | null;
-}
-
-interface DateRangeInput {
-  startDate?: string | undefined;
-  endDate?: string | undefined;
-}
-
-interface BlankRateInput extends DateRangeInput {
-  wakeHour: number;
-  sleepHour: number;
-}
-
-interface TimePLInput {
-  startDate: string;
-  endDate: string;
-  prevStart?: string | undefined;
-  prevEnd?: string | undefined;
-  visibleDateKeys?: readonly string[] | undefined;
-  prevVisibleDateKeys?: readonly string[] | undefined;
-  wakeHour: number;
-  sleepHour: number;
-}
-
-interface StatsPageDataInput {
-  startDate: string;
-  endDate: string;
-  prevStart: string;
-  prevEnd: string;
-  visibleDateKeys?: readonly string[] | undefined;
-  prevVisibleDateKeys?: readonly string[] | undefined;
-  year: number;
-  monthlyMonths: number;
-  wakeHour: number;
-  sleepHour: number;
-}
-
-interface TagDashboardInput {
-  tagId: string;
-  startDate: string;
-  endDate: string;
-  limit: number;
-}
-
-function roundTo1(value: number): number {
-  return Math.round(value * 10) / 10;
-}
-
-function filterRowsByVisibleDateKeys<T extends { start_at: string }>(
-  rows: readonly T[],
-  visibleDateKeys: readonly string[] | undefined,
-  timezone: string,
-): T[] {
-  if (!visibleDateKeys) return [...rows];
-
-  const visibleDateKeySet = new Set(visibleDateKeys);
-  return rows.filter((row) =>
-    visibleDateKeySet.has(formatInTimeZone(new Date(row.start_at), timezone, 'yyyy-MM-dd')),
-  );
-}
+import type { DateRangeInput } from './statistics-fetchers';
+import { StatisticsGeneralService } from './statistics-general-service';
+import type { BlankRateInput } from './statistics-kpi-service';
+import { StatisticsKpiService } from './statistics-kpi-service';
+import type { StatsPageDataInput, TimePLInput } from './statistics-summary-service';
+import { StatisticsSummaryService } from './statistics-summary-service';
+import type { TagDashboardInput } from './statistics-tag-dashboard-service';
+import { StatisticsTagDashboardService } from './statistics-tag-dashboard-service';
+import type { ServiceSupabaseClient } from './types';
 
 export class StatisticsService {
-  constructor(private readonly supabase: ServiceSupabaseClient) {}
+  private readonly generalService: StatisticsGeneralService;
+  private readonly kpiService: StatisticsKpiService;
+  private readonly summaryService: StatisticsSummaryService;
+  private readonly tagDashboardService: StatisticsTagDashboardService;
+
+  constructor(supabase: ServiceSupabaseClient) {
+    this.generalService = new StatisticsGeneralService(supabase);
+    this.kpiService = new StatisticsKpiService(supabase);
+    this.summaryService = new StatisticsSummaryService(supabase, this.kpiService);
+    this.tagDashboardService = new StatisticsTagDashboardService(supabase);
+  }
 
   // ---------------------------------------------------------------------------
   // General: タグ別統計・時間帯分布
   // ---------------------------------------------------------------------------
 
   /** `get_tag_stats` 相当。実績（records）ベースのタグ別件数・最終使用日。 */
-  async getTagStats(userId: string): Promise<{
-    counts: Record<string, number>;
-    lastUsed: Record<string, string>;
-  }> {
-    const records = await this.fetchRecords(userId);
-    const byTag = new Map<string, { count: number; lastUsed: string | null }>();
-    for (const record of records) {
-      if (record.tag_id == null) continue;
-      const acc = byTag.get(record.tag_id) ?? { count: 0, lastUsed: null };
-      acc.count += 1;
-      if (acc.lastUsed == null || record.start_at > acc.lastUsed) acc.lastUsed = record.start_at;
-      byTag.set(record.tag_id, acc);
-    }
-
-    const rows = Array.from(byTag.entries()).map(([tag_id, v]) => ({
-      tag_id,
-      record_count: v.count,
-      last_used: v.lastUsed,
-    }));
-    return aggregateTagStats(rows);
+  async getTagStats(userId: string) {
+    return this.generalService.getTagStats(userId);
   }
 
   /** `get_time_by_tag` 相当。実績（records）のタグ別合計時間。 */
   async getTimeByTag(userId: string, range: DateRangeInput = {}) {
-    const [records, tagsById] = await Promise.all([
-      this.fetchRecords(userId, range),
-      this.fetchTagsById(userId),
-    ]);
-
-    const minutesByTag = new Map<string, number>();
-    for (const record of records) {
-      if (record.tag_id == null) continue;
-      minutesByTag.set(
-        record.tag_id,
-        (minutesByTag.get(record.tag_id) ?? 0) + minutesBetween(record.start_at, record.end_at),
-      );
-    }
-
-    const rows = Array.from(minutesByTag.entries())
-      .filter(([, minutes]) => minutes > 0)
-      .map(([tagId, minutes]) => {
-        const tag = tagsById.get(tagId);
-        return {
-          tag_id: tagId,
-          tag_name: tag?.name ?? '',
-          tag_color: tag?.color ?? 'indigo',
-          hours: minutes / 60,
-        };
-      })
-      .sort((a, b) => b.hours - a.hours);
-
-    return transformTimeByTagResponse(rows);
+    return this.generalService.getTimeByTag(userId, range);
   }
 
   /** `get_daily_hours` 相当。指定年の日別実績時間（ヒートマップ用）。 */
   async getDailyHours(userId: string, year: number) {
-    const timezone = await getUserTimezone(this.supabase, userId);
-    const startOfYear = fromZonedTime(`${year}-01-01T00:00:00`, timezone).toISOString();
-    const startOfNextYear = fromZonedTime(`${year + 1}-01-01T00:00:00`, timezone).toISOString();
-    const records = await this.fetchRecords(userId, {
-      startDate: startOfYear,
-      endDate: startOfNextYear,
-    });
-    return groupHoursByDay(records, timezone);
+    return this.generalService.getDailyHours(userId, year);
   }
 
   /** `get_hourly_distribution` 相当。 */
   async getHourlyDistribution(userId: string, range: DateRangeInput = {}) {
-    const timezone = await getUserTimezone(this.supabase, userId);
-    const records = await this.fetchRecords(userId, range);
-    return aggregateHourlyDistribution(groupMinutesByHour(records, timezone));
+    return this.generalService.getHourlyDistribution(userId, range);
   }
 
   /** `get_dow_distribution` 相当。 */
   async getDayOfWeekDistribution(userId: string, range: DateRangeInput = {}) {
-    const timezone = await getUserTimezone(this.supabase, userId);
-    const records = await this.fetchRecords(userId, range);
-    return aggregateDayOfWeekDistribution(groupMinutesByDow(records, timezone));
+    return this.generalService.getDayOfWeekDistribution(userId, range);
   }
 
   /** `get_monthly_hours` 相当。 */
   async getMonthlyTrend(userId: string, months = 12) {
-    const timezone = await getUserTimezone(this.supabase, userId);
-    const nowStr = formatInTimeZone(new Date(), timezone, 'yyyy-MM');
-    const [nowYear, nowMonth] = nowStr.split('-').map(Number) as [number, number];
-    const startDate = getMonthlyStartDate(nowYear, nowMonth, months);
-
-    const records = await this.fetchRecords(userId, { startDate: startDate.toISOString() });
-    return aggregateMonthlyTrend(groupHoursByMonth(records, timezone), nowYear, nowMonth, months);
+    return this.generalService.getMonthlyTrend(userId, months);
   }
 
   // ---------------------------------------------------------------------------
@@ -238,22 +84,12 @@ export class StatisticsService {
    * 詳細は `domain/estimation-accuracy.ts` の `aggregatePlanRecordEstimationAccuracy` を参照。
    */
   async getEstimationAccuracy(userId: string, range: DateRangeInput = {}) {
-    const [plans, tagsById] = await Promise.all([
-      this.fetchPlans(userId, range),
-      this.fetchTagsById(userId),
-    ]);
-    const rows = await this.computeEstimationAccuracy(userId, plans, tagsById);
-    return transformEstimationAccuracy(rows);
+    return this.kpiService.getEstimationAccuracy(userId, range);
   }
 
   /** `get_blank_rate` 相当。予定（plans）ベースのスケジュール時間から空白率を算出する。 */
-  async getBlankRate(userId: string, { startDate, endDate, wakeHour, sleepHour }: BlankRateInput) {
-    const plans = await this.fetchPlans(userId, { startDate, endDate });
-    const scheduledMinutes = plans.reduce(
-      (sum, plan) => sum + minutesBetween(plan.start_at, plan.end_at),
-      0,
-    );
-    return computeBlankRate(scheduledMinutes, { startDate, endDate, wakeHour, sleepHour });
+  async getBlankRate(userId: string, input: BlankRateInput) {
+    return this.kpiService.getBlankRate(userId, input);
   }
 
   // ---------------------------------------------------------------------------
@@ -261,503 +97,30 @@ export class StatisticsService {
   // ---------------------------------------------------------------------------
 
   /** `get_active_dates` 相当。実績（records）が存在する日付（tz basis）の一覧。 */
-  async getActiveDates(userId: string, startDate: string): Promise<string[]> {
-    const timezone = await getUserTimezone(this.supabase, userId);
-    const records = await this.fetchRecords(userId, { startDate });
-    const days = new Set(
-      records.map((record) => formatInTimeZone(new Date(record.start_at), timezone, 'yyyy-MM-dd')),
-    );
-    return Array.from(days).sort();
+  async getActiveDates(userId: string, startDate: string) {
+    return this.summaryService.getActiveDates(userId, startDate);
   }
 
   /** `get_stats_kpi_summary` 相当。 */
-  async getStatsOverview(
-    userId: string,
-    { startDate, endDate, wakeHour, sleepHour }: BlankRateInput,
-  ) {
-    const timezone = await getUserTimezone(this.supabase, userId);
-    const [records, plans] = await Promise.all([
-      this.fetchRecords(userId, { startDate, endDate }),
-      this.fetchPlans(userId, { startDate, endDate }),
-    ]);
-
-    const cumulativeMinutes = records.reduce(
-      (sum, record) => sum + minutesBetween(record.start_at, record.end_at),
-      0,
-    );
-    const plannedEntries = records.filter((record) => record.plan_id != null).length;
-    const contextSwitches = computeContextSwitches(records, timezone);
-    const scheduledMinutes = plans.reduce(
-      (sum, plan) => sum + minutesBetween(plan.start_at, plan.end_at),
-      0,
-    );
-    const blankRate = computeBlankRate(scheduledMinutes, {
-      startDate,
-      endDate,
-      wakeHour,
-      sleepHour,
-    });
-
-    return transformStatsOverviewResponse({
-      cumulativeTime: { totalMinutes: cumulativeMinutes },
-      planRate: {
-        totalEntries: records.length,
-        plannedEntries,
-        planRate: records.length > 0 ? plannedEntries / records.length : 0,
-      },
-      contextSwitches,
-      blankRate,
-    });
+  async getStatsOverview(userId: string, input: BlankRateInput) {
+    return this.summaryService.getStatsOverview(userId, input);
   }
 
   /** `get_time_pl_data` 相当。タグ別の予実（budget/actual）+ 日次ポイント。 */
-  async getTimePLData(userId: string, input: TimePLInput): Promise<TimePLResponse> {
-    const {
-      startDate,
-      endDate,
-      prevStart,
-      prevEnd,
-      visibleDateKeys,
-      prevVisibleDateKeys,
-      wakeHour,
-      sleepHour,
-    } = input;
-    const timezone = await getUserTimezone(this.supabase, userId);
-
-    const [rangePlans, rangeRecords, tagsById] = await Promise.all([
-      this.fetchPlans(userId, { startDate, endDate }),
-      this.fetchRecords(userId, { startDate, endDate }),
-      this.fetchTagsById(userId),
-    ]);
-    const plans = filterRowsByVisibleDateKeys(rangePlans, visibleDateKeys, timezone);
-    const records = filterRowsByVisibleDateKeys(rangeRecords, visibleDateKeys, timezone);
-    const tags = this.buildTagPL(plans, records, tagsById);
-
-    let prevTags: TimePLResponse['prevTags'] = [];
-    if (prevStart && prevEnd) {
-      const [rangePrevPlans, rangePrevRecords] = await Promise.all([
-        this.fetchPlans(userId, { startDate: prevStart, endDate: prevEnd }),
-        this.fetchRecords(userId, { startDate: prevStart, endDate: prevEnd }),
-      ]);
-      const prevPlans = filterRowsByVisibleDateKeys(rangePrevPlans, prevVisibleDateKeys, timezone);
-      const prevRecords = filterRowsByVisibleDateKeys(
-        rangePrevRecords,
-        prevVisibleDateKeys,
-        timezone,
-      );
-      prevTags = this.buildTagPL(prevPlans, prevRecords, tagsById);
-    }
-
-    const availableMinutes = computeAvailableMinutesInclusive({
-      startDate,
-      endDate,
-      wakeHour,
-      sleepHour,
-      timezone,
-      visibleDayCount: visibleDateKeys?.length,
-    });
-
-    return { tags, prevTags, availableMinutes };
+  async getTimePLData(userId: string, input: TimePLInput) {
+    return this.summaryService.getTimePLData(userId, input);
   }
 
   /** `get_stats_page_data` 相当。Review panel 用データを一括構築する。 */
-  async getStatsPageData(userId: string, input: StatsPageDataInput): Promise<StatsPageData> {
-    const {
-      startDate,
-      endDate,
-      prevStart,
-      prevEnd,
-      visibleDateKeys,
-      prevVisibleDateKeys,
-      year,
-      monthlyMonths,
-      wakeHour,
-      sleepHour,
-    } = input;
-    const timezone = await getUserTimezone(this.supabase, userId);
-
-    const startOfYear = fromZonedTime(`${year}-01-01T00:00:00`, timezone).toISOString();
-    const startOfNextYear = fromZonedTime(`${year + 1}-01-01T00:00:00`, timezone).toISOString();
-    const nowStr = formatInTimeZone(new Date(), timezone, 'yyyy-MM');
-    const [nowYear, nowMonth] = nowStr.split('-').map(Number) as [number, number];
-    const monthlyStartDate = getMonthlyStartDate(nowYear, nowMonth, monthlyMonths);
-
-    const [
-      rangeRecords,
-      rangePlans,
-      rangePrevRecords,
-      rangePrevPlans,
-      yearRecords,
-      monthlyRecords,
-      tagsById,
-    ] = await Promise.all([
-      this.fetchRecords(userId, { startDate, endDate }),
-      this.fetchPlans(userId, { startDate, endDate }),
-      this.fetchRecords(userId, { startDate: prevStart, endDate: prevEnd }),
-      this.fetchPlans(userId, { startDate: prevStart, endDate: prevEnd }),
-      this.fetchRecords(userId, { startDate: startOfYear, endDate: startOfNextYear }),
-      this.fetchRecords(userId, { startDate: monthlyStartDate.toISOString() }),
-      this.fetchTagsById(userId),
-    ]);
-    const records = filterRowsByVisibleDateKeys(rangeRecords, visibleDateKeys, timezone);
-    const plans = filterRowsByVisibleDateKeys(rangePlans, visibleDateKeys, timezone);
-    const prevRecords = filterRowsByVisibleDateKeys(
-      rangePrevRecords,
-      prevVisibleDateKeys,
-      timezone,
-    );
-    const prevPlans = filterRowsByVisibleDateKeys(rangePrevPlans, prevVisibleDateKeys, timezone);
-
-    const overview = this.buildOverviewSection(records);
-    const prevOverview = this.buildOverviewSection(prevRecords);
-    const contextSwitches = computeContextSwitches(records, timezone);
-    const scheduledMinutes = plans.reduce(
-      (sum, plan) => sum + minutesBetween(plan.start_at, plan.end_at),
-      0,
-    );
-    const fullBlankRate = computeBlankRate(scheduledMinutes, {
-      startDate,
-      endDate,
-      wakeHour,
-      sleepHour,
-      visibleDayCount: visibleDateKeys?.length,
-    });
-
-    const timeByTag = transformTimeByTagResponse(this.buildTimeByTagRows(records, tagsById));
-    // NOTE: get_stats_page_data の hourly/dow は 24h/7dow の raw bucket をそのまま返す
-    // （getHourlyDistribution/getDayOfWeekDistribution procedure が行う 2h slot 集約や
-    // 曜日ラベル変換は適用しない。RPC 契約どおり）。
-    const hourly = groupMinutesByHour(records, timezone).map((row) => ({
-      hour: row.hour,
-      totalMinutes: row.total_minutes,
-    }));
-    const dow = groupMinutesByDow(records, timezone).map((row) => ({
-      dow: row.dow,
-      totalMinutes: row.total_minutes,
-    }));
-    const energyMap = groupEnergyMap(records, timezone);
-    const prevEnergyMap = groupEnergyMap(prevRecords, timezone);
-
-    const estimationAccuracy = transformEstimationAccuracy(
-      await this.computeEstimationAccuracy(userId, plans, tagsById),
-    );
-    const prevEstimationAccuracy = transformEstimationAccuracy(
-      await this.computeEstimationAccuracy(userId, prevPlans, tagsById),
-    );
-
-    const dailyHours = groupHoursByDay(yearRecords, timezone);
-    const monthlyTrend = aggregateMonthlyTrend(
-      groupHoursByMonth(monthlyRecords, timezone),
-      nowYear,
-      nowMonth,
-      monthlyMonths,
-    );
-
-    return {
-      overview,
-      prevOverview,
-      contextSwitches,
-      blankRate: {
-        availableMinutes: fullBlankRate.availableMinutes,
-        scheduledMinutes: fullBlankRate.scheduledMinutes,
-        blankRate: fullBlankRate.blankRate,
-      },
-      timeByTag,
-      hourly,
-      dow,
-      energyMap,
-      estimationAccuracy,
-      prevEstimationAccuracy,
-      prevEnergyMap,
-      dailyHours,
-      monthlyTrend,
-    };
+  async getStatsPageData(userId: string, input: StatsPageDataInput) {
+    return this.summaryService.getStatsPageData(userId, input);
   }
 
   // ---------------------------------------------------------------------------
   // タグ詳細ダッシュボード（旧 `tagStatisticsRouter.getTagDashboard`）
   // ---------------------------------------------------------------------------
 
-  async getTagDashboard(userId: string, { tagId, startDate, endDate, limit }: TagDashboardInput) {
-    const timezone = await getUserTimezone(this.supabase, userId);
-    const [tag, records, plans] = await Promise.all([
-      this.fetchTagById(userId, tagId),
-      this.fetchRecordsOverlapping(userId, tagId, startDate, endDate),
-      this.fetchPlansOverlapping(userId, tagId, startDate, endDate),
-    ]);
-
-    const plansById = new Map(plans.map((plan) => [plan.id, plan]));
-
-    // 1 plan に複数 record（分割記録）が紐づく場合、予定時間の二重計上を避けるため
-    // 「代表 record」1 件だけに planned range を割り当てる。from_plan があればそれを優先する。
-    const primaryRecordIdByPlanId = new Map<string, string>();
-    for (const record of records) {
-      if (record.plan_id == null) continue;
-      const currentId = primaryRecordIdByPlanId.get(record.plan_id);
-      if (!currentId) {
-        primaryRecordIdByPlanId.set(record.plan_id, record.id);
-        continue;
-      }
-      if (record.source === 'from_plan') {
-        primaryRecordIdByPlanId.set(record.plan_id, record.id);
-      }
-    }
-
-    const rows: TagDashboardTimeblockRow[] = records.map((record) => {
-      const plan = record.plan_id ? plansById.get(record.plan_id) : undefined;
-      const isPrimary =
-        record.plan_id != null && primaryRecordIdByPlanId.get(record.plan_id) === record.id;
-      return {
-        id: record.id,
-        title: record.title,
-        description: record.note,
-        start_time: isPrimary && plan ? plan.start_at : null,
-        end_time: isPrimary && plan ? plan.end_at : null,
-        actual_start_time: record.start_at,
-        actual_end_time: record.end_at,
-        tag_id: record.tag_id,
-      };
-    });
-
-    // 未記録・未 skip の plan は「予定のみ」の行として残す（実績時間は 0）。
-    const recordedPlanIds = new Set(
-      records.map((record) => record.plan_id).filter((id): id is string => id != null),
-    );
-    for (const plan of plans) {
-      if (recordedPlanIds.has(plan.id) || plan.skipped_at) continue;
-      rows.push({
-        id: plan.id,
-        title: plan.title,
-        description: plan.note,
-        start_time: plan.start_at,
-        end_time: plan.end_at,
-        actual_start_time: null,
-        actual_end_time: null,
-        tag_id: plan.tag_id,
-      });
-    }
-
-    return buildTagDashboard({ tag, rows, limit, timezone });
-  }
-
-  // ---------------------------------------------------------------------------
-  // private: 見積もり精度の共通計算
-  // ---------------------------------------------------------------------------
-
-  private async computeEstimationAccuracy(
-    userId: string,
-    plans: ReadonlyArray<StatPlanRow>,
-    tagsById: ReadonlyMap<string, TagLookupRow>,
-  ): Promise<EstimationAccuracyDbRow[]> {
-    const planIds = plans.map((plan) => plan.id);
-    const records = planIds.length > 0 ? await this.fetchRecordsByPlanIds(userId, planIds) : [];
-
-    const planRows = plans
-      .filter((plan): plan is StatPlanRow & { tag_id: string } => plan.tag_id != null)
-      .map((plan) => ({
-        id: plan.id,
-        tag_id: plan.tag_id,
-        planned_minutes: minutesBetween(plan.start_at, plan.end_at),
-      }));
-    const recordRows = records.map((record) => ({
-      plan_id: record.plan_id,
-      source: record.source,
-      minutes: minutesBetween(record.start_at, record.end_at),
-    }));
-    const tagLookup: Map<string, EstimationAccuracyTagLookup> = new Map(
-      Array.from(tagsById.entries()).map(([id, tag]) => [id, { name: tag.name, color: tag.color }]),
-    );
-
-    return aggregatePlanRecordEstimationAccuracy(planRows, recordRows, tagLookup);
-  }
-
-  private buildOverviewSection(records: ReadonlyArray<StatRecordRow>) {
-    const totalMinutes = records.reduce(
-      (sum, record) => sum + minutesBetween(record.start_at, record.end_at),
-      0,
-    );
-    const recordCount = records.length;
-    const totalEntries = records.length;
-    const plannedEntries = records.filter((record) => record.plan_id != null).length;
-    return {
-      totalMinutes,
-      recordCount,
-      totalEntries,
-      plannedEntries,
-      planRate: totalEntries > 0 ? plannedEntries / totalEntries : 0,
-    };
-  }
-
-  private buildTimeByTagRows(
-    records: ReadonlyArray<StatRecordRow>,
-    tagsById: ReadonlyMap<string, TagLookupRow>,
-  ) {
-    const minutesByTag = new Map<string, number>();
-    for (const record of records) {
-      if (record.tag_id == null) continue;
-      minutesByTag.set(
-        record.tag_id,
-        (minutesByTag.get(record.tag_id) ?? 0) + minutesBetween(record.start_at, record.end_at),
-      );
-    }
-    return Array.from(minutesByTag.entries())
-      .filter(([, minutes]) => minutes > 0)
-      .map(([tagId, minutes]) => {
-        const tag = tagsById.get(tagId);
-        return {
-          tag_id: tagId,
-          tag_name: tag?.name ?? '',
-          tag_color: tag?.color ?? 'indigo',
-          hours: minutes / 60,
-        };
-      })
-      .sort((a, b) => b.hours - a.hours);
-  }
-
-  private buildTagPL(
-    plans: ReadonlyArray<StatPlanRow>,
-    records: ReadonlyArray<StatRecordRow>,
-    tagsById: ReadonlyMap<string, TagLookupRow>,
-  ): TimePLResponse['tags'] {
-    interface Accumulator {
-      budget: number;
-      actual: number;
-      hasPlan: boolean;
-    }
-    const byTag = new Map<string, Accumulator>();
-
-    for (const plan of plans) {
-      if (plan.tag_id == null) continue;
-      const acc = byTag.get(plan.tag_id) ?? { budget: 0, actual: 0, hasPlan: false };
-      acc.budget += minutesBetween(plan.start_at, plan.end_at);
-      acc.hasPlan = true;
-      byTag.set(plan.tag_id, acc);
-    }
-    for (const record of records) {
-      if (record.tag_id == null) continue;
-      const acc = byTag.get(record.tag_id) ?? { budget: 0, actual: 0, hasPlan: false };
-      acc.actual += minutesBetween(record.start_at, record.end_at);
-      byTag.set(record.tag_id, acc);
-    }
-
-    return Array.from(byTag.entries())
-      .filter(([, acc]) => acc.budget > 0 || acc.actual > 0)
-      .map(([tagId, acc]) => {
-        const tag = tagsById.get(tagId);
-        return {
-          tagId,
-          tagName: tag?.name ?? '',
-          tagColor: tag?.color ?? 'indigo',
-          tagIcon: tag?.icon ?? null,
-          budgetMinutes: roundTo1(acc.budget),
-          actualMinutes: roundTo1(acc.actual),
-          isPlanned: acc.hasPlan,
-        };
-      })
-      .sort((a, b) => b.actualMinutes - a.actualMinutes);
-  }
-
-  // ---------------------------------------------------------------------------
-  // private: fetch
-  // ---------------------------------------------------------------------------
-
-  private async fetchRecords(userId: string, range: DateRangeInput = {}): Promise<StatRecordRow[]> {
-    let query = this.supabase
-      .from(databaseTables.records)
-      .select('id, tag_id, plan_id, source, start_at, end_at')
-      .eq('user_id', userId)
-      .is('deleted_at', null);
-    if (range.startDate) query = query.gte('start_at', range.startDate);
-    if (range.endDate) query = query.lt('start_at', range.endDate);
-
-    const { data, error } = await query;
-    if (error) throw new Error(`Failed to fetch records: ${error.message}`);
-    return data ?? [];
-  }
-
-  private async fetchRecordsByPlanIds(userId: string, planIds: string[]): Promise<StatRecordRow[]> {
-    const { data, error } = await this.supabase
-      .from(databaseTables.records)
-      .select('id, tag_id, plan_id, source, start_at, end_at')
-      .eq('user_id', userId)
-      .is('deleted_at', null)
-      .in('plan_id', planIds);
-    if (error) throw new Error(`Failed to fetch records by plan ids: ${error.message}`);
-    return data ?? [];
-  }
-
-  private async fetchPlans(userId: string, range: DateRangeInput = {}): Promise<StatPlanRow[]> {
-    let query = this.supabase
-      .from('plans')
-      .select('id, tag_id, start_at, end_at')
-      .eq('user_id', userId)
-      .is('deleted_at', null);
-    if (range.startDate) query = query.gte('start_at', range.startDate);
-    if (range.endDate) query = query.lt('start_at', range.endDate);
-
-    const { data, error } = await query;
-    if (error) throw new Error(`Failed to fetch plans: ${error.message}`);
-    return data ?? [];
-  }
-
-  private async fetchTagsById(userId: string): Promise<Map<string, TagLookupRow>> {
-    const { data, error } = await this.supabase
-      .from('tags')
-      .select('id, name, color, icon')
-      .eq('user_id', userId);
-    if (error) throw new Error(`Failed to fetch tags: ${error.message}`);
-    return new Map((data ?? []).map((tag) => [tag.id, tag]));
-  }
-
-  private async fetchTagById(userId: string, tagId: string): Promise<TagDashboardTagRow> {
-    const { data, error } = await this.supabase
-      .from('tags')
-      .select('id, name, color, icon')
-      .eq('user_id', userId)
-      .eq('id', tagId)
-      .eq('is_active', true)
-      .single();
-    if (error || !data) throw new Error(`Tag not found: ${error?.message ?? tagId}`);
-    return data;
-  }
-
-  private async fetchRecordsOverlapping(
-    userId: string,
-    tagId: string,
-    startDate: string,
-    endDate: string,
-  ): Promise<Array<StatRecordRow & { title: string; note: string | null }>> {
-    const { data, error } = await this.supabase
-      .from(databaseTables.records)
-      .select('id, title, note, tag_id, plan_id, source, start_at, end_at')
-      .eq('user_id', userId)
-      .eq('tag_id', tagId)
-      .is('deleted_at', null)
-      .lt('start_at', endDate)
-      .gt('end_at', startDate)
-      .order('start_at', { ascending: true });
-    if (error) throw new Error(`Failed to fetch records for tag dashboard: ${error.message}`);
-    return data ?? [];
-  }
-
-  private async fetchPlansOverlapping(
-    userId: string,
-    tagId: string,
-    startDate: string,
-    endDate: string,
-  ): Promise<
-    Array<StatPlanRow & { title: string; note: string | null; skipped_at: string | null }>
-  > {
-    const { data, error } = await this.supabase
-      .from('plans')
-      .select('id, title, note, tag_id, start_at, end_at, skipped_at')
-      .eq('user_id', userId)
-      .eq('tag_id', tagId)
-      .is('deleted_at', null)
-      .lt('start_at', endDate)
-      .gt('end_at', startDate)
-      .order('start_at', { ascending: true });
-    if (error) throw new Error(`Failed to fetch plans for tag dashboard: ${error.message}`);
-    return data ?? [];
+  async getTagDashboard(userId: string, input: TagDashboardInput) {
+    return this.tagDashboardService.getTagDashboard(userId, input);
   }
 }
